+++ packages/@groupby/elements-sayt-plugin/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] [minor]
+### Added
+- ELE-247: Added `SaytOptions` interface.
+
+### Changed
+- ELE-247: Changed options intake interface from the imported `SaytConfig` to the extended `SaytOptions`.
+
 ## [0.1.0] - 2019-11-28
 ### Added
 - SFX-156: Added the `SaytPlugin` module.