+++ packages/@groupby/elements-sayt-plugin/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] [patch]
+## [0.1.1] - 2019-12-19
 ### Changed
 - ELE-247: Changed the type of the `SaytPlugin` options object to be `SaytPluginOptions` instead of `SaytConfig`.